separated hammer behaviour from rigid body
