+++ Assets/Scripts/Player/Hammer/VisualHammer.cs
@@ -1,0 +1,30 @@
+using UnityEngine;
+
+namespace Hammer
+{
+    public class VisualHammer : MonoBehaviour
+    {
+        [SerializeField] private Transform pivotTransform;
+        private Rigidbody _rb;
+
+        void Awake()
+        {
+            _rb = GetComponent<Rigidbody>();
+        }
+
+        void Update()
+        {
+
+        }
+
+        // this should probably be replaced with damage on the enemies or something (it should also involve force maybe?)
+        public void OnCollisionEnter(Collision collision)
+        {
+            if (collision.gameObject.CompareTag("Enemy"))
+            {
+                Destroy(collision.gameObject);
+            }
+        }
+    }
+
+}
